# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "business-project API",
    "description": "Контракт между бэкендом и клиентами",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/2fa/disable": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "disable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/2fa/enable": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "enable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpEnableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/2fa/setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "setup",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpSetupResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/2fa/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "verify_2fa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify2faRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/auth/email": {
      "post": {
        "tags": [
          "account"
        ],
        "operationId": "request_email_change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "письмо с подтверждением отправлено"
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/email/confirm": {
      "post": {
        "tags": [
          "account"
        ],
        "operationId": "confirm_email_change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/password": {
      "post": {
        "tags": [
          "account"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "пароль изменён, все сессии закрыты"
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/verify-email": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "verify_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChangeEmailRequest": {
        "type": "object",
        "required": [
          "new_email",
          "current_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "ConfirmEmailRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "Credentials": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "Тело ответа с ошибкой — единственный формат на весь API.\n`fields` появляется только у ошибок про конкретные поля формы.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "машиночитаемый код (он же ключ локализации)"
          },
          "fields": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "message": {
            "type": "string",
            "description": "текст на языке запроса — можно показывать пользователю как есть"
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string",
            "description": "имя поля в теле запроса, например \"password\""
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LoginResponse": {
        "type": "object",
        "required": [
          "requires_2fa"
        ],
        "properties": {
          "access_token": {
            "type": [
              "string",
              "null"
            ],
            "description": "присутствует, если 2FA не требуется"
          },
          "pending_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "requires_2fa": {
            "type": "boolean",
            "description": "если true — второй шаг: POST /api/auth/2fa/verify с pending_token"
          },
          "user": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ]
      },
...
      "TokenResponse": {
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "TotpCodeRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "TotpEnableRequest": {
        "type": "object",
        "required": [
          "secret",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      },
      "TotpSetupResponse": {
        "type": "object",
        "required": [
          "secret",
          "otpauth_url"
        ],
        "properties": {
          "otpauth_url": {
            "type": "string",
            "description": "ссылка otpauth:// для приложения-аутентификатора"
          },
          "secret": {
            "type": "string",
            "description": "показать пользователю для ручного ввода"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "totp_enabled",
          "email_verified"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "как обращаться к пользователю; не задано — показываем почту"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean",
            "description": "подтверждена ли почта (по ссылке из письма)"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "язык интерфейса и писем: код из i18n::ALL_LANGS"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "totp_enabled": {
            "type": "boolean",
            "description": "включён ли второй фактор"
          }
        }
      },
      "Verify2faRequest": {
        "type": "object",
        "required": [
          "pending_token",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "pending_token": {
            "type": "string"
          }
        }
      },
      "VerifyEmailRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}